backend: RPC/nb: Remove request argument from unmarshal_data
This RPC is called directly, that is, without being wrapped by the run()
function, because it needs to run in the same kernel with the open
notebook. Therefore, there isn't a request object to pass.

If we want to log something inside this RPC in the future, then we
should reconsider its arguments.

Signed-off-by: Ilias Katsakioris <elikatsis@arrikto.com>