backend: Improve dependencies detection algorithm (#83)
* Fix a regression introduced in #81

During the detection of all the candidates to be marshalled from a step,
functions and classes were skipped. In this way steps could not marshal
functions or classes. This commit fixes that.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* Rename variable to avoid potential conflicts

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* Save node's graph info at the end of iteration

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* Search for chained deps inside fns with free vars

Functions can be dependent on variables outside of the scope the
function body (free variables). These free variables can be themselves
functions that in turn use other free variables. This chain of
dependecies can be arbitrary long. This commit covers this case by
iteratively searching for the available functions in a step.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* Update tests

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* Correct parsing of functions with a return stmt

By parsing the function def statement first and just then parsing the
body detecting the missing name with pyflakes, the report would not
detect missing names on the `return` line, becase the report would just
say that it was outside of a function.

This commit solves that, by simplifying how functions are parsed, since
having individual arguments was not needed anyway.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>