Add annotation and limit tags and GPU UI (#101)
* backend: Parse cell annotations and limits tags

Add 2 new tags to the Kale tags:

- limit: these tags specify K8s resource limits to be applied to a
step's pod. The syntax is `limit:<limit_key>:<limit_value>`.
- annotation: these tags work similarly to `limit`, but specify
annotations to be applied to the step's pod.

These tags are expected in the notebook cells metadata, alongside the
`step` and `prev` tags. Moreover, they are expected just in a cell that
also has a `step` tag. This is will be true in general, since the UI
will enforce this behavior.

* backend: Set limits and annotations to ContainerOp

Use the KFP SDK to add annotations and limits to the steps.

* labextension: Add a cell metadata dialog

Introduce a new dialog to set GPU limits to steps. This dialog is just a
first iteration of what will be a richer form to add any sort of K8s
resource limits, annotations and other useful information to compile the
pipeline.

The implementation is pretty much specific to GPU field, but will be
reworked to be more dynamic in the future.

* labextension: Handle limit tags in TagsUtils

* labextension: Add an inline limit tags summary

This implementation is specific for GPU resource limits, as this is the
only kind of limit that is supported. As more limits and cell tags are
added, the generation of this summary will be refactored to be more
dynamic.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>