backend: Handle exiting from IPython gracefully
Running the function `sys.exit()` raises a special exception `SystemExit`
that is not caught by the python REPL, making it exit the program.
IPython's REPL, instead, does catch `SystemExit`. It displays the message
and then goes back to the REPL.

Code that could either run in an IPython kernel (because the Kale pipeline
was produced from a notebook) or in a standard Python process, needs to
handle the exit process seamlessly, regardless of where it's running.

If a Kale function wants to terminate the Python process and it could be
running either inside IPython or a standard Python process, then it
should use `utils.graceful_exit`. In case the code is running inside
inside IPython, `graceful_exit` will raise a `KaleGracefulExit`
exception that is expected to be captured by
`kale.utils.jupyter_utils.capture_streams`.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>