logging: Add support for logging empty lines
Using a standard logger, it's not possible to just print empty lines
without having to print the Formatter prefix as well. This can be a
limitation when we want to use just the logger for user-facing outputs,
but we also want to be able to space out some section, producing empty
lines.

This commit extends the base Logger class with a method that, when
called, it temporarily suppress all the handlers' formatters with the
empty string, logs empty lines, and then restores the original
formatters.

NOTE: This function is not thread safe. If multiple processes are using
the same logger concurrently, there might be cases when one thread uses
the logger while its formatters are being suppressed by another thread.
Currently, Kale is not multithreaded, so this is not an issue.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>