marshal: Support tf.keras model serialization
To support the tf.keras model serialization a new backend is added
to save the mode in the SavedModel format. This produces a folder
instead of a single file. Thus, utils.py file is also altered to
see folders as serialized objects in the case of TensorFlow artifacts.

The choice to proceed with SavedModel format instead of a single .h5 file
was taken because SavedModel is the new standard as of TF 2.x.
Later, we can use this format as is to deploy models as web services,
to edge devices or TensorFlow JS.

Signed-off-by: Dimitris Poulopoulos <dimpo@arrikto.com>
Reviewed-by: Stefano Fioravanzo <stefano@arrikto.com>