marshal: Set compile to False when loading model
The model will load in inference mode and won't be ready for further
training. If the user wants to train it further he would have to
compile it again. However, this is preferable in order to support
custom loss functions and optimizers.

For example, if the user has trained a model with a custom loss
function or a custom optimizer and tries to load it with compile set
to True a `ValueError` will occur specifying that it does not recognize
this loss function or optimizer. Thus the step will fail on KFP.

Signed-off-by: Dimitris Poulopoulos <dimpo@arrikto.com>
Reviewed-by: Stefano Fioravanzo <stefano@arrikto.com>