backend/mlmd: Fix logic of finding ancestors
To find a step's ancestors, we have been parsing the workflow status and
checking for nodes which have the step as a child. Then we took these
parents and tried to `get()` them.
This does not work when such parent is not of Pod type (for example,
when the parent node is a Retry node).

With this commit we change the logic of retrieving ancestors to find all
the step's parents that are of Pod type.

Signed-off-by: Ilias Katsakioris <elikatsis@arrikto.com>
Reviewed-by: Stefano Fioravanzo <stefano@arrikto.com>