marshal: Allow a backend to exit on missing lib
The default behavior of marshal backends is to try to pickle an object
with dill in case the specialized library of a custom backend is not
resolved. This is done to cover some (potential) cases when the object
can still be marshalled using `dill`.

In the case of SKLearn object, we don't want this fallback to happen
because SKLearn object **can** be saved with `dill` (`pickle`), so if the
user didn't install `joblib` (which is not a dependency of `sklearn`,
they would end up saving the model with `pickle`.

This poses a problem for when the marshal backend is used for serving,
since KFServing requires SKLearn object to be saved with `joblib`.

This commit thus introduces a general way for custom backend to opt out
from this default fallback, and applied it to the SKLearn backned.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>