sdk: Add artifact decorator
This commit introduces a new SDK decorator that allows users to declare
the artifacts created during a step execution. The `@artifact` decorator
is thus supposed to be used on functions that have already been
decorated with the `@step` decorator.

Declaring step artifacts must be done before the step actually executes,
because Kale needs to create a pipeline that is aware of these outputs,
and properly update each step's output artifacts dictionary in the KFP
DSL.

Currently only HTML artifacts are supported.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>