backend: Improve base docker image detection
Just reading the Pod spec's container image is not enough to have a
reproducible reference to the current image, because an image tag can be
re-assigned to newer builds, in the future (e.g. when using the `latest`
tag). The only way to have reproducible reference is by using the
image manifest's `sha`.

Kubernetes exposes this in the Pod's `status`, under `containerStatuses`
[1], in the field `imageID`. In the case this field is empty (this could
happen when the image was built locally), then fallback to reading the
Pod's container `image` field.

[1] https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.20/#containerstatus-v1-core  # noqa: 501

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>