backend: Fix pip 20.3 dependency resolution
With the new dependency resolver introduced in pip 20.3, dependency
resolution would get stuck because of the following conflicts:

1. `kfserving` requires `kubernetes==10.0.1`
2. Kale's `setup.py` specifies `kubernetes>=11.0.0`

We introduced (2) in [1]. This commit reverts that change and manually
catches the `TypeError` exception.

`kfserving` version 0.5.0 changes the `kubernetes` requirement to
install versions >=12.0.0. However, this is incompatible with KFP
requirement (<12.0.0).
This is also incompatible with our code, since in version 12.0.0 there
are breaking changes on code we import (e.g., Exceptions).

As a result, we change the restrictions as follows:
* `kfserving >= 0.4.0, < 0.5.0`
* `kubernetes < 12.0.0`

Closes https://github.com/kubeflow-kale/kale/issues/273
Closes https://github.com/kubeflow-kale/kale/issues/275

[1] https://github.com/kubeflow-kale/kale/commit/e3bc079a6e971e42b02aa04fcfaceb4ea28faf31

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>